backend/requirements.txt
asn1crypto==0.23.0
attrs==17.2.0
Automat==0.6.0
certifi==2017.7.27.1
cffi==1.11.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.0.3
cssselect==1.0.1
Django==1.11.5
django-filter==1.0.4
djangorestframework==3.6.4
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
lxml==4.0.0
Markdown==2.6.9
parsel==1.2.0
pyasn1==0.3.6
pyasn1-modules==0.1.4
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==17.3.0
pytz==2017.2
queuelib==1.4.2
requests==2.18.4
Scrapy==1.4.0
service-identity==17.0.0
six==1.11.0
Twisted==17.9.0
urllib3==1.22
w3lib==1.18.0
zope.interface==4.4.3
